Added Logic and Quality on CoreSavings
