+++ app/src/main/AndroidManifest.xml
@@ -23,31 +23,31 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".CoreSavings"
+            android:name=".coresavings.CoreSavings"
             android:exported="true" />
         <activity
-            android:name=".CoreBudget"
+            android:name=".coresavings.CoreBudget"
             android:exported="true" />
         <activity
-            android:name=".CoreBudgetInput"
+            android:name=".coresavings.CoreBudgetInput"
             android:exported="true" />
         <activity
-            android:name=".CoreBudgetEdit"
+            android:name=".coresavings.CoreBudgetEdit"
             android:exported="true" />
         <activity
-            android:name=".CoreBudgetDetail"
+            android:name=".coresavings.CoreBudgetDetail"
             android:exported="true" />
         <activity
-            android:name=".CoreGoal"
+            android:name=".coresavings.CoreGoal"
             android:exported="true" />
         <activity
-            android:name=".CoreGoalInput"
+            android:name=".coresavings.CoreGoalInput"
             android:exported="true" />
         <activity
-            android:name=".CoreGoalEdit"
+            android:name=".coresavings.CoreGoalEdit"
             android:exported="true" />
         <activity
-            android:name=".CoreGoalDetail"
+            android:name=".coresavings.CoreGoalDetail"
             android:exported="true" />
         <activity
             android:name=".MainActivity"
