+++ app/src/main/java/com/example/corewallet/coresavings/CoreGoalDetail.kt
@@ -1,13 +1,9 @@
-package com.example.corewallet
+package com.example.corewallet.coresavings
 
 import android.graphics.Color
 import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Button
 import androidx.activity.ComponentActivity
+import com.example.corewallet.R
 
 
 class CoreGoalDetail : ComponentActivity() {
