+++ app/src/main/res/layout/core_budget.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CoreBudget">
+    tools:context=".coresavings.CoreBudget">
 
     <LinearLayout
         android:layout_width="match_parent"
