+++ app/src/main/res/layout/core_budget_detail.xml
@@ -5,7 +5,7 @@
     android:id="@+id/coreGoal"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CoreGoal">
+    tools:context=".coresavings.CoreGoal">
 
     <LinearLayout
         android:layout_width="match_parent"
