+++ app/src/main/res/layout/core_budget_edit.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CoreBudgetInput">
+    tools:context=".coresavings.CoreBudgetInput">
 
     <LinearLayout
         android:layout_width="match_parent"
