+++ app/src/main/res/layout/core_goal_edit.xml
@@ -4,7 +4,7 @@
     android:id="@+id/coreGoal"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CoreGoalEdit">
+    tools:context=".coresavings.CoreGoalEdit">
 
     <LinearLayout
         android:layout_width="match_parent"
